Refactor: abstract equationfieldset
