+++ client/src/views/inspector/equation/EditFlowEquationFormView.tsx
@@ -1,66 +1,9 @@
 import { getParentEntities } from "@/actions/graphTraversal";
 import { Badge } from "@/components/ui/badge";
-import { Field, FieldLabel, FieldSet } from "@/components/ui/field";
-import { Input } from "@/components/ui/input";
 import { useFlowForm } from "@/controllers/form";
-import type { Flow, Node, Stock } from "@/models/graph";
-import { useGraphStore } from "@/stores/graph";
-import {
-    buildLabelToIdMap,
-    removeInvalidCharacters,
-    replaceEquationIdsWithLabels,
-    replaceEquationLabelsWithIds,
-} from "@/utils/equation";
 
+import { EquationFieldSet } from "./EquationFieldSet";
 import { EquationFormWrapper } from "./EquationFormWrapper";
-
-function FlowEquationFieldSet({
-    flowId,
-    parents,
-}: {
-    flowId: string;
-    parents: Array<Node | Stock | Flow>;
-}) {
-    const { flow, handleChange } = useFlowForm(flowId);
-    if (!flow) {
-        return null;
-    }
-
-    const labelMap = buildLabelToIdMap(parents);
-    const state = useGraphStore.getState();
-
-    return (
-        <FieldSet>
-            <Field>
-                <FieldLabel>Equation</FieldLabel>
-                <Input
-                    name="equation"
-                    value={replaceEquationIdsWithLabels(
-                        flow.equation,
-                        state.nodes,
-                        state.stocks,
-                        state.flows,
-                    )}
-                    onChange={(e) => {
-                        handleChange({
-                            equation: replaceEquationLabelsWithIds(e.target.value, labelMap),
-                        });
-                    }}
-                    onBlur={() => {
-                        const validEquation = removeInvalidCharacters(
-                            flow.equation,
-                            state.nodes,
-                            state.flows,
-                            state.stocks,
-                        );
-
-                        handleChange({ equation: validEquation });
-                    }}
-                />
-            </Field>
-        </FieldSet>
-    );
-}
 
 export function EditFlowEquationFormView({ flowId }: { flowId: string }) {
     const { flow, handleCancel, handleChange } = useFlowForm(flowId);
@@ -77,7 +20,7 @@ export function EditFlowEquationFormView({ flowId }: { flowId: string }) {
             onDelete={() => handleChange({ equation: "" })}
             showDelete
         >
-            <FlowEquationFieldSet flowId={flowId} parents={parents} />
+            <EquationFieldSet entity={flow} handleChange={handleChange} parents={parents} />
             {parents.length > 0 && (
                 <div className="flex flex-wrap gap-2">
                     {parents?.map((parent) => {
